# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.Attachment:
    properties:
      createdAt:
        type: string
      fileName:
        type: string
      filePath:
        type: string
      fileSize:
        type: integer
      id:
        type: string
      messageId:
        type: string
      mimeType:
        type: string
      uploadedBy:
        type: string
      workspaceId:
        type: string
    type: object
  models.AttachmentUpload:
    properties:
      fileName:
        type: string
      fileSize:
        type: integer
      id:
        type: string
      isImage:
        type: boolean
      mimeType:
        type: string
      preview:
        description: Base64 preview for images
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Multi-Agent Collaboration Platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Orchestra API
  version: "1.0"
paths:
...
  /api/workspaces/{id}/attachments:
    get:
      description: List all attachments in a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by conversation ID
        in: query
        name: conversationId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Attachment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List attachments
      tags:
      - attachments
  /api/workspaces/{id}/attachments/{attachmentId}:
    delete:
      description: Delete a file attachment
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachmentId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete attachment
      tags:
      - attachments
    get:
      description: Download a file attachment
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachmentId
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download attachment
      tags:
      - attachments
  /api/workspaces/{id}/attachments/{attachmentId}/info:
    get:
      description: Get attachment metadata
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachmentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Attachment'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get attachment info
      tags:
      - attachments
...
  /api/workspaces/{id}/conversations/{convId}/attachments:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file attachment to a conversation
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Conversation ID
        in: path
        name: convId
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Sender member ID
        in: formData
        name: senderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AttachmentUpload'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload attachment
      tags:
      - attachments
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"